% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_prices.R
\docType{data}
\name{crypto_prices}
\alias{crypto_prices}
\title{Prices of 3 crypto currencies between 2016-04-01 and 2019-07-01}
\format{A data frame with 1174 rows and 4 variables:
\itemize{
\item \emph{Date}: date when the price was recorded
\item \emph{LTC}: closing price of Litecoin
\item \emph{BTC}: closing price of Bitconin
\item \emph{EHT}: closing price of Ethereum
}}
\usage{
crypto_prices
}
\description{
A dataset closing prices for Litecoin, Bitcoin and Ethereum
on 1174 days. Prices are recorded in US dollars.
}
\keyword{datasets}
